backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
boto3
pandas
numpy
joblib
xgboost
scikit-learn
cloudinary
python-dotenv
supabase
passlib[bcrypt]
tensorflow
shap>=0.44.0

asyncpg

databases
